Start Of Interpreter!
This is the start of the actual project! Today I got through the creation of the TokenType enum to catagories the lexemes in our streams, next time ill continue with the token class as well as the actual scanning for our project! The Token object is be in red at the moment without the actual class built, so that worried me for a bit that I had already done something wrong!