Tackled the Token class
made sense of the Token processing, and the steps of scanning without implementation. learned about the case processing of scanning which makes sense to me, and how we are building up and creating tokens for processing through literally checking every single character of our file which makes sense! This is were we could use regex as a solution for sure, but we want to do it ourselves! Also this is so cool because its literally the regular language that the coding language IS. Scanning is that actual manifestation of theory, were you get to build out what terms, functions and layouts of expression are done at least for now character wise in the system! Thats so cool!