Staring on ML and AST
Finally started working again! So interseting to learn how this all works, we are now in the land between the parser and the interpreter, converting our tokens into stuff that is more reconisiable to a computer! we have looked at grammar notation which I willl make a mermaid graph of to represent our slighly higher language that tokens!